Style Refactoring
