+++ projeto-vida/src/components/LandPage/BannerHome/Style/BannerHome.scss
