+++ projeto-vida/src/components/LandPage/BannerWarranty/Style/BannerWarrantyResponsive.scss
