+++ projeto-vida/src/components/LandPage/Courses/Style/CardCourse.scss
