+++ projeto-vida/src/components/LandPage/Courses/Style/CardCourseResponsive.scss
