+++ projeto-vida/src/components/LandPage/Courses/Style/Courses.scss
