+++ projeto-vida/src/components/LandPage/Courses/Style/CoursesResponsive.scss
