+++ projeto-vida/src/components/LandPage/EventSection/Style/EventCard.scss
