+++ projeto-vida/src/components/LandPage/EventSection/Style/EventCardResponsive.scss
